requirements.txt
Django==2.1.2
django-debug-toolbar==2.0
django-mysql==3.2.0
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
jsmin==2.2.2
mysql-connector==2.2.9
mysql-connector-python==8.0.17
Pillow==6.1.0
protobuf==3.9.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
